# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API for Gopher",
    "title": "Gopher",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/v1",
  "paths": {
    "/authentication/token": {
      "post": {
        "description": "Creates a token for a user",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["authentication"],
        "summary": "Creates a token for a user (login)",
        "parameters": [
          {
            "description": "User credentials",
            "name": "payload",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/main.LoginUserPayload"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JWT token for authentication",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {}
          },
          "401": {
            "description": "Unauthorized",
            "schema": {}
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {}
          }
        }
      }
    },
    "/authentication/user": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Register a new user",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["authentication"],
        "summary": "Registers a user",
        "parameters": [
          {
            "description": "User credentials",
            "name": "payload",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/main.RegisterUserPayload"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User registered",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {}
          },
          "404": {
            "description": "Not Found",
            "schema": {}
          }
        }
      }
    },
    "/notes/{professorID}": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Creates a note where the user can upload files",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["notes"],
        "summary": "Creates a note",
        "parameters": [
          {
            "type": "integer",
            "description": "professor ID",
            "name": "professorID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/main.CreateNotePayload"
            }
          }
        }
      }
    },
    "/school/{schoolID}": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Gets a school by ID",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["schools"],
        "summary": "Gets a school",
        "parameters": [
          {
            "type": "integer",
            "description": "School ID",
            "name": "SchoolID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "good",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {}
          },
          "404": {
            "description": "Not Found",
            "schema": {}
          }
        }
      }
    },
    "/search/school": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Gets a school by its name",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["search"],
        "summary": "Gets all schools by similar name",
        "parameters": [
          {
            "type": "string",
            "description": "School Name",
            "name": "q",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {}
          },
          "404": {
            "description": "Not Found",
            "schema": {}
          }
        }
      }
    },
    "/users/activate/{token}": {
      "put": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Activates/Register a user by invitation token",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["users"],
        "summary": "Activates/Register a user",
        "parameters": [
          {
            "type": "string",
            "description": "Invitation token",
            "name": "token",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "User activated",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {}
          },
          "404": {
            "description": "Not Found",
            "schema": {}
          }
        }
      }
    },
    "/users/{userID}": {
      "get": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "description": "Fetches a user profile by ID",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["users"],
        "summary": "Fetches a user profile",
        "parameters": [
          {
            "type": "integer",
            "description": "User ID",
            "name": "userID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/store.User"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "main.CreateNotePayload": {
      "type": "object",
      "required": ["content", "subject", "title"],
      "properties": {
        "content": {
          "type": "string"
        },
        "files_url": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "professor_id": {
          "type": "integer"
        },
        "subject": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "main.LoginUserPayload": {
      "type": "object",
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string",
          "maxLength": 255
        },
        "password": {
          "type": "string",
          "maxLength": 72
        }
      }
    },
    "main.RegisterUserPayload": {
      "type": "object",
      "required": ["email", "password", "username"],
      "properties": {
        "email": {
          "type": "string",
          "maxLength": 255
        },
        "password": {
          "type": "string",
          "maxLength": 72,
          "minLength": 6
        },
        "username": {
          "type": "string",
          "maxLength": 60
        }
      }
    },
    "store.Role": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "level": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "store.User": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "is_active": {
          "type": "boolean"
        },
        "role": {
          "$ref": "#/definitions/store.Role"
        },
        "role_id": {
          "type": "integer"
        },
        "username": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}